feat: clean up for new integration with module establishment (#21)
