--- backend/src/restaurants/enums/restaurant-type.enum.ts
@@ -1,9 +1,0 @@
-export enum RestaurantType {
-  RESTAURANT = 'RESTAURANT',
-  CATERER = 'CATERER',
-  FAST_FOOD = 'FAST_FOOD',
-  FOOD_TRUCK = 'FOOD_TRUCK',
-  BAR = 'BAR',
-  COLLECTIVE_CATERING = 'COLLECTIVE_CATERING',
-  OTHER = 'OTHER',
-}
