--- backend/src/restaurants/restaurants.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { RestaurantsService } from './restaurants.service';
-import { RestaurantsController } from './restaurants.controller';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Restaurant } from './entities/restaurant.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Restaurant])],
-  controllers: [RestaurantsController],
-  providers: [RestaurantsService],
-})
-export class RestaurantsModule {}
