--- backend/src/restaurants/restaurants.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { RestaurantsService } from './restaurants.service';
-
-describe('RestaurantsService', () => {
-  let service: RestaurantsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [RestaurantsService],
-    }).compile();
-
-    service = module.get<RestaurantsService>(RestaurantsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
